Simplified project
Moving project generators to another repo, this module will ONLY parse swagger docs & render templates
