+++ .vscode/launch.json
@@ -8,7 +8,7 @@
       "type": "node",
       "request": "launch",
       "name": "Launch Program",
-      "program": "${workspaceFolder}\\lib\\codegen.mjs"
+      "program": "${workspaceFolder}\\example\\example.js"
     }
   ]
 }
