--- demoSwaggerModule.mjs
@@ -1,10 +1,0 @@
-import demo from './demo';
-
-const outputPath = './dist/api/controllers/pet.mjs';
-
-const imports = [
-  { name: 'util', path: 'util' },
-  { name: 'db', path: '../lib/db' },
-];
-
-demo('./templates/swaggerModule.mustache', outputPath, imports);
