+++ example/controllerGenerator.mustache
