--- lib/codegen.mjs
@@ -1,58 +1,0 @@
-import mustache from 'mustache';
-import _ from 'lodash';
-import swaggerTools from 'swagger-tools';
-
-const spec = swaggerTools.specs.v2;
-
-function processSwagger(moduleName, swagger, imports, constants) {
-  const data = {
-    moduleName,
-    imports,
-    constants,
-    domain: (swagger.schemes && swagger.schemes.length > 0 && swagger.host && swagger.basePath) ? `${swagger.schemes[0]}://${swagger.host}${swagger.basePath.replace(/\/+$/g, '')}` : '',
-    info: swagger.info,
-    classes: swagger.tags,
-    operations: [],
-    definitions: [],
-    securityDefinitions: [],
-  };
-
-  _.forEach(swagger.paths, (methods, path) => {
-    _.forEach(methods, (operation, method) => {
-      operation.path = path;
-      operation.method = method;
-      data.operations.push(operation);
-    });
-  });
-
-  _.forEach(swagger.definitions, (definition, name) => {
-    definition.name = name;
-    data.definitions.push(definition);
-  });
-
-  _.forEach(swagger.securityDefinitions, (definition, name) => {
-    definition.name = name;
-    data.securityDefinitions.push(definition);
-  });
-
-  return data;
-}
-
-function getCode({
-  moduleName,
-  template,
-  swagger,
-  imports,
-  constants,
-}) {
-  if (!template || !swagger || !moduleName) {
-    throw new Error('Missing required parameter.');
-  }
-  const data = processSwagger(moduleName, swagger, imports, constants);
-  return mustache.render(
-    template,
-    data,
-  );
-}
-
-export default getCode;
