+++ package.json
@@ -23,11 +23,13 @@
   "license": "ISC",
   "homepage": "https://github.com/garrettsutula/swagger-js-codegen",
   "dependencies": {
-    "commander": "^2.9.0",
-    "js-yaml": "^3.10.0",
     "lodash": "^4.17.10",
     "mustache": "2.3.0",
     "swagger-tools": "^0.10.3"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.12.0"
+  }
 }
