--- templates/node/cjs/swaggerApp.mustache
@@ -1,23 +1,0 @@
-require express = 'express';
-require swaggerTools = 'swagger-tools';
-{{#imports}}
-require {{name}} from './{{name}}';
-{{/imports}}
-
-const swaggerObject = require('../api/swagger/swagger.json');
-const app = express();
-module.exports = app;
-
-var routerOptions = {
-  swaggerUi: swaggerObject,
-  controllers: './api/controllers'
-};
-
-swaggerTools.initializeMiddleware(swaggerObject, (middleware) => {
-  app.use(middleware.swaggerMetadata());
-  app.use(middleware.swaggerValidator());
-  app.use(middleware.swaggerRouter(routerOptions));
-  app.use(middleware.swaggerUi());
-  const port = {{constants.port}} || 8080;
-  app.listen(port);
-});
