--- templates/node/cjs/swaggerController.mustache
@@ -1,34 +1,0 @@
-{{#imports}}
-import {{name}} from '{{path}}';
-{{/imports}}
-{{#classes}}
-import {{name}} from './{{name}}';
-{{/classes}}
-
-{{#operations}}
-function {{operationId}}(req, res) {
-  {{#parameters}}
-  const {{name}} = req.swagger.params['{{name}}'].value;
-  {{/parameters}}
-  {{#tags}}
-  {{.}}.{{operationId}}(
-    {{#parameters}}
-    {{name}},
-    {{/parameters}}
-  )
-    .then((response) => {
-      res.send(response);
-    })
-    .catch((err) => {
-      console.log(err.message);
-      res.send(err.message);
-    });
-  {{/tags}}
-}
-
-{{/operations}}
-export default {
-  {{#operations}}
-  {{operationId}},
-  {{/operations}}
-};
