--- templates/node/ejs/swaggerApp.mustache
@@ -1,22 +1,0 @@
-{{#imports}}
-import {{name}} from '{{path}}';
-{{/imports}}
-
-const app = express();
-export default app;
-
-const swaggerUi = swaggerUiDist.absolutePath();
-app.use(express.static(swaggerUi));
-
-const config = {
-  appRoot: path.dirname(), // required config
-};
-
-SwaggerExpress.create(config, (err, swaggerExpress) => {
-  if (err) { throw err; }
-
-  // install middleware
-  swaggerExpress.register(app);
-  const port = {{constants.port}};
-  app.listen(port);
-});
