--- templates/swaggerPackage.mustache
@@ -1,22 +1,0 @@
-{
-  "name": "{{info.title}}",
-  "version": "{{info.version}}",
-  "description": "{{info.description}}",
-  "keywords": [],
-  "author": "{{info.author}}",
-  "license": "{{info.license.name}}",
-  "main": "app.mjs",
-  "dependencies": {
-    {{#imports}}
-    "{{name}}": "{{version}}",
-    {{/imports}}
-  },
-  "devDependencies": {
-    "should": "^7.1.0",
-    "supertest": "^1.0.0"
-  },
-  "scripts": {
-    "start": "node app.js",
-    "test": "swagger project test"
-  }
-}
